requirements.txt
streamlit==1.37.1
pydantic==2.8.2
python-dotenv==1.0.1
PyPDF2==3.0.1
tiktoken==0.7.0
openai==1.43.0